pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kubera-intel"
version = "0.1.0"
description = "Financial data collection and AI-powered analysis tool"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "SanGyuk-Raccoon", email = "dmstjkim80@gmail.com"}]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-genai>=1.0.0",
]

[project.scripts]
kubera-intel = "intel.cli:main"

[project.urls]
Homepage = "https://github.com/oh-my-kubera/kubera-intel"
Repository = "https://github.com/oh-my-kubera/kubera-intel"
Issues = "https://github.com/oh-my-kubera/kubera-intel/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/intel"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
